pyproject.toml
[project]
name = "sales-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.6",
    "asyncpg>=0.30.0",
    "polars>=1.33.1",
    "psycopg2>=2.9.10",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "flask>=3.0.3",
    "requests>=2.32.3",
]
